Added formatting util
